backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
stripe==7.0.0
python-dotenv==1.0.0
Pillow>=10.2.0
cryptography==41.0.7
djongo==1.3.6  # MongoDB adapter for Django ORM
pymongo==3.12.1  # MongoDB Python driver (compatible with djongo 1.3.6)
sqlparse==0.2.4  # Required by djongo (note: Django 4.2.7 prefers >=0.3.1, but djongo requires 0.2.4)
setuptools>=65.0.0  # Required for pkg_resources

